OPCT-11: fix counters on plugin report progress (#42)
https://issues.redhat.com/browse/OPCT-11

Bump to v0.3.0-alpha1 after plugin fix
https://github.com/redhat-openshift-ecosystem/provider-certification-plugins/pull/34

This PR will improve the counters when running OPCT, fixing a bug on
openshift-tests on 4.12+ clusters.

We don't need to create a alpha release for CLI, but having this change
on the master will get newer results from OpenShift CI, running weekly
for 4.12.