OPCT-9: introduce the report cmd to automate the review steps (#22)
This PR introduces the command `report` to be used on the result review
stage (post-submit to Red Hat), applying different filters according to
checks on the Baseline results and CI flake tests (using Sippy API as a
data source)

This command also dumps the failed filtered tests into text files,
extracting the details for each failed test to be used while
troubleshooting the post-execution.

The following files are optional files to be used alongside the archive:

- baseline results : reference artifacts execution in the OPCT CI
Pipeline to help to review the providers artifact

Example of CLI
```bash
./openshift-provider-cert-linux-amd64 report \
  --baseline ./opct_baseline-ocp_4.11.4-platform_none-aws-202210102258_sonobuoy_7a895e01-3d3e-44cc-a5d3-ac4f2ed678fd.tar.gz \
  202210132151_sonobuoy_6af99324-2dc6-4de4-938c-200b84111481.tar.gz  \
  --save-to ./processed-results
```

There are spikes to understand if we can collect include/collect it on
the certification runtime, so we can allow end-users to use this command
before submitting their results.

https://issues.redhat.com/browse/OPCT-9
https://issues.redhat.com/browse/OPCT-10