OPCT-239: create basic changelog builder based in commits (#74)
Generate changelog by commits when mkdocs is building the docs.

The reason for this change is to improve the report for each release,
currently, there is a GitHub action that generates a delta when a new
release is created but the delta is always based on the last GitHub
release, not the last Major version, most interested by us as the
project uses SemVer standards.