OPCT-226: cmd/status: Adding flag to set the watch interval (#87)
The interval watching, and printing in the stdout is not customized
leading to a high amount of messages in the stdout - it is good for
troubleshooting purpose, but a downside is it increases the amount of
logs stored in CI with actionable information.

This PR introduces a possibility to set the flag `--watch-interval=int`
to the `status` command, also called by the `run` command.

This change is isolated from the `report` enhancement covered in #76