OPCT-257: Doc - remove machineset option in dedicated node section (#89)
Remove form the user guide the ambiguous instruction to setup dedicated
node as we are requiring an agnostic installation, and providing an
option to use MachineSet, but MAPI isn't available in agnostic
installations/platform type None.

This option can be valid when we'll create different guides. Currently
the "User Guide" is very specific for VCSP requirements. The intention
of #88 is to improve it, but we don't have ETA yet.