OPCT-308: introduce check OPCT-032E to validate install-time
Introducing the check OPCT-032E to validate the total time since the
cluster was installed and conformance tests had been executed.

This is a common issue when partners installing a cluster, runs many
conformance tests executions, or just run the final conformance in old
clusters which logs and metrics have been contaminated with
non-conformance tests executions, making the framework to raise
false-positive signals.

This check is satisfying the diagram introduced in:
https://github.com/redhat-openshift-ecosystem/opct/pull/168
https://redhat-openshift-ecosystem.github.io/opct/guides/cluster-validation/#process-overview

which guides the partner to open a ticket when executed in a fresh
cluster.